src/backend/pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.1.0"
description = "FastAPI backend for Todo webapp"
authors = ["Tabraiz Haider"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.29"
psycopg = {extras = ["binary"], version = "^3.1.18"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.3.4"
email-validator = "^2.2.0"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"
ruff = "^0.4.6"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
